Big refactor for clarifying the code on the canyon_observer crate. Upgrade on the code documentation
