+++ canyon_observer/src/lib.rs
@@ -1,11 +1,14 @@
 pub mod credentials;
 pub mod handler;
 mod memory;
+mod constants;
+
+// Database Engine related
+pub mod postgresql;
 
 extern crate canyon_crud;
 
-use handler::CanyonRegisterEntity;
-
+use postgresql::register_types::CanyonRegisterEntity;
 // use credentials::DatabaseCredentials;
 
 /// Holds the data needed by Canyon when the host
@@ -17,7 +20,6 @@ use handler::CanyonRegisterEntity;
 /// Also takes care about track what data structures Canyon
 /// should be managing
 pub static mut QUERIES_TO_EXECUTE: Vec<String> = Vec::new();
-// pub static mut CANYON_REGISTER_ENTITIES: Vec<CanyonEntity> = Vec::new();
 pub static mut CANYON_REGISTER_ENTITIES: Vec<CanyonRegisterEntity> = Vec::new();
-// pub static REGISTER: *const Vec<CanyonEntity> = unsafe { &CANYON_REGISTER_ENTITIES as *const Vec<CanyonEntity> };
-// pub static mut CREDENTIALS: Option<DatabaseCredentials> = None;
+
+// TODO Change it for lazy_static! elements
