+++ canyon_observer/src/postgresql/mod.rs
@@ -1,0 +1,3 @@
+pub mod information_schema;
+pub mod migrations;
+pub mod register_types;
